Export and document useChartWidth, useChartHeight
